requirements.txt
google-adk>=0.1.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
qdrant-client>=1.7.0
litellm>=1.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
fakeredis>=2.20.0
httpx>=0.27.0
tenacity>=8.0.0
structlog>=24.0.0
upstash_redis>=2.0.0
